Fix(🛠): Incorrect data display for completed and inactive bar on instructor dashboard.

+++ classes/Instructor.php
@@ -526,7 +526,7 @@ class Instructor {
 				continue;
 			}
 
-			$course_progress = tutor_utils()->get_course_completed_percent( $enrollment->post_parent, $enrollment->post_author );
+			$course_progress = (int) tutor_utils()->get_course_completed_percent( $enrollment->post_parent, $enrollment->post_author );
 
 			if ( 100 === $course_progress ) {
 				++$counts['completed'];
